requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pypdf==5.1.0
scikit-learn==1.5.2
scipy==1.14.1
joblib==1.4.2
requests==2.32.3
